last

+++ 0x05-pointers_arrays_strings/101-keygen.c
@@ -10,9 +10,10 @@
 
 int main(void)
 {
-	srand((unsigned int)(time(NULL)));
+	srand(time(NULL));
 
 	int index = 0;
+	
 	char characters[] = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789/,.-+=~`<>:";
 	for(index = 0; index < 12; index++)
 	{
